+++ man-roxygen/return-tsc-api-response.R
@@ -1,0 +1,11 @@
+#' @return An S3 object of class 'tsc_api_response' containing:
+#'
+#'   data: The parsed response content
+#'
+#'   serializer: The called serializer, e.g. 'supra'
+#'
+#'   url: The called URL
+#'
+#'   date: The timestamp of the request
+#'
+#'   status_code: The HTTP status code of the request