+++ tests/testthat/test-gs_getWFS.R
@@ -1,0 +1,8 @@
+test_that("gs_getWFS returns a list", {
+  # skip if KMI GeoServer is offline or unreachable
+  x <- gs_getWFS()
+  expect_equal(class(x), "list")
+})
+
+
+# usethis::use_r("gs_getWFS")